Fix premature 'free' call in producer function.
Previously, the producer thread would free the buffer for the line
after it had added it to the queue. However, since the consumer still
needs to reference this string, this results in a use-after-free
error. The fix for this is to free the string after its use in the
consumer thread.